requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
websocket-client==1.6.4
